add icons and fix types
