+++ package.json
@@ -22,6 +22,7 @@
     "@trpc/server": "next",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "lucide-react": "^0.378.0",
     "next": "^14.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
