+++ src/app/_components/icon.tsx
@@ -1,0 +1,15 @@
+import { icons } from "lucide-react";
+
+interface IconProps {
+  name: keyof typeof icons;
+  color?: string;
+  size?: number | string;
+}
+
+const Icon: React.FC<IconProps> = ({ name, color, size }) => {
+  const LucideIcon = icons[name];
+
+  return <LucideIcon color={color} size={size} />;
+};
+
+export default Icon;
