+++ src/server/api/routers/airtable.ts
@@ -1,14 +1,15 @@
 import { createTRPCRouter, publicProcedure } from "~/server/api/trpc";
 import { TRPCError } from "@trpc/server";
+import { AirtableResponse, Projects } from "~/types"; // Import Projects
 
-// Correctly set these in your .env.local file
 const BASE_ID = process.env.AIRTABLE_BASE_ID;
 const TABLE_ID = process.env.AIRTABLE_TABLE_ID;
 const API_KEY = process.env.AIRTABLE_PAT;
 const AIRTABLE_URL = `https://api.airtable.com/v0/${BASE_ID}/${TABLE_ID}`;
 
 export const airtableRouter = createTRPCRouter({
-  fetchRecords: publicProcedure.query(async () => {
+  fetchRecords: publicProcedure.query<Projects>(async () => {
+    // Use Projects as the return type
     try {
       const response = await fetch(AIRTABLE_URL, {
         headers: {
@@ -24,8 +25,8 @@ export const airtableRouter = createTRPCRouter({
         });
       }
 
-      const data = await response.json();
-      return data.records;
+      const data = (await response.json()) as AirtableResponse;
+      return data.records; // This now correctly matches the Projects type
     } catch (error) {
       throw new TRPCError({
         code: "INTERNAL_SERVER_ERROR",
